Add property-based tests for GcsSinkTask in group-by-key mode
